requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.2.8
django-extensions==4.1
django-filter==25.2
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
gprof2dot==2025.4.14
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mysqlclient==2.2.7
pillow==12.0.0
pydot==4.0.1
PyJWT==2.10.1
pyparsing==3.2.5
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
